% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlidingWindows.R
\name{SlidingWindows}
\alias{SlidingWindows}
\title{Sliding Windows.}
\usage{
SlidingWindows(y, w)
}
\arguments{
\item{y}{A vector containing univariate time series.}

\item{w}{An integer value indicating the window size \eqn{w < length(y)}.
If \eqn{w = length(y)}, will be computed the function will not slide.}
}
\value{
A list containing "w", "SlidingWindows".
}
\description{
This function generates sliding windows approach of a time series.
}
\details{
This function return the matrix with time series sliding windows.
}
\examples{
y <- rnorm(100)
SlidingWindows(y,w=99)

}
\references{
Guedes, E.F. Modelo computacional para análise de movimentos e co-movimentos de mercados financeiros, Ph.D. thesis, Programa de Pós-graduação em Modelagem Computacional e Tecnologia Industrial. Centro Universitário Senai Cimatec, 2019.
}
